refactor: verification service 통합
