+++ src/main/java/com/planetrush/planetrush/verification/service/VerificationServiceImpl.java
@@ -11,6 +11,7 @@ import com.planetrush.planetrush.planet.exception.PlanetNotFoundException;
 import com.planetrush.planetrush.planet.repository.PlanetRepository;
 import com.planetrush.planetrush.verification.domain.VerificationRecord;
 import com.planetrush.planetrush.verification.repository.VerificationRecordRepository;
+import com.planetrush.planetrush.verification.repository.custom.VerificationRecordRepositoryCustom;
 import com.planetrush.planetrush.verification.service.dto.VerificationResultDto;
 
 import lombok.RequiredArgsConstructor;
@@ -23,6 +24,7 @@ public class VerificationServiceImpl implements VerificationService {
 	private final MemberRepository memberRepository;
 	private final PlanetRepository planetRepository;
 	private final VerificationRecordRepository verificationRecordRepository;
+	private final VerificationRecordRepositoryCustom verificationRecordRepositoryCustom;
 
 	/**
 	 * {@inheritDoc}
@@ -52,4 +54,16 @@ public class VerificationServiceImpl implements VerificationService {
 			.build());
 	}
 
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public boolean getTodayRecord(Long memberId, Long planetId) {
+		Member member = memberRepository.findById(memberId)
+			.orElseThrow(() -> new MemberNotFoundException("Member not found with ID: " + memberId));
+		Planet planet = planetRepository.findById(planetId)
+			.orElseThrow(() -> new PlanetNotFoundException("Planet not found with ID: " + planetId));
+		VerificationRecord verificationRecord = verificationRecordRepositoryCustom.findTodayRecord(member, planet);
+		return verificationRecord != null;
+	}
 }
